+++ mail/mail-uniffi/CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased] - 2024-00-00
 
+### Changed
+
+ - Opening a Draft now always syncs the contents from the server.
+
 ## [0.27.0] - 2024-12-17
 
 ### Added